+++ cli/tests/test_git_init.rs
@@ -222,7 +222,7 @@ fn test_git_init_external(bare: bool) {
         ");
         // Non-colocated repo shouldn't record the Git HEAD
         insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-        Workspace is currently not colocated with Git.
+        Workspace 'default' is currently not colocated with Git.
         Last imported/exported Git HEAD: (none)
         [EOF]
         ");
@@ -479,7 +479,7 @@ fn test_git_init_colocated_via_git_repo_path() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -494,7 +494,7 @@ fn test_git_init_colocated_via_git_repo_path() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: f3fe58bc88ccfb820b930a21297d8e48bf76ac2a
     [EOF]
     ");
@@ -528,7 +528,7 @@ fn test_git_init_colocated_via_git_repo_path_gitlink() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -543,7 +543,7 @@ fn test_git_init_colocated_via_git_repo_path_gitlink() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: f3fe58bc88ccfb820b930a21297d8e48bf76ac2a
     [EOF]
     ");
@@ -576,7 +576,7 @@ fn test_git_init_colocated_via_git_repo_path_symlink_directory() -> TestResult {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -591,7 +591,7 @@ fn test_git_init_colocated_via_git_repo_path_symlink_directory() -> TestResult {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: f3fe58bc88ccfb820b930a21297d8e48bf76ac2a
     [EOF]
     ");
@@ -630,7 +630,7 @@ fn test_git_init_colocated_via_git_repo_path_symlink_directory_without_bare_conf
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -645,7 +645,7 @@ fn test_git_init_colocated_via_git_repo_path_symlink_directory_without_bare_conf
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: f3fe58bc88ccfb820b930a21297d8e48bf76ac2a
     [EOF]
     ");
@@ -686,7 +686,7 @@ fn test_git_init_colocated_via_git_repo_path_symlink_gitlink() -> TestResult {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -701,7 +701,7 @@ fn test_git_init_colocated_via_git_repo_path_symlink_gitlink() -> TestResult {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&jj_work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: f3fe58bc88ccfb820b930a21297d8e48bf76ac2a
     [EOF]
     ");
@@ -928,7 +928,7 @@ fn test_git_init_external_but_git_dir_exists() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently not colocated with Git.
+    Workspace 'default' is currently not colocated with Git.
     Last imported/exported Git HEAD: (none)
     [EOF]
     ");
@@ -957,7 +957,7 @@ fn test_git_init_colocated_via_flag_git_dir_exists() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -972,7 +972,7 @@ fn test_git_init_colocated_via_flag_git_dir_exists() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: f3fe58bc88ccfb820b930a21297d8e48bf76ac2a
     [EOF]
     ");
@@ -1002,7 +1002,7 @@ fn test_git_init_colocated_via_config_git_dir_exists() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -1017,7 +1017,7 @@ fn test_git_init_colocated_via_config_git_dir_exists() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: f3fe58bc88ccfb820b930a21297d8e48bf76ac2a
     [EOF]
     ");
@@ -1080,7 +1080,7 @@ fn test_git_init_colocated_via_flag_overrides_false_config() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");
@@ -1104,7 +1104,7 @@ fn test_git_init_colocated_via_flag_git_dir_not_exists() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: (none)
     [EOF]
     ");
@@ -1122,7 +1122,7 @@ fn test_git_init_colocated_via_flag_git_dir_not_exists() {
     [EOF]
     ");
     insta::assert_snapshot!(get_colocation_status(&work_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: (none)
     [EOF]
     ");
@@ -1276,7 +1276,7 @@ fn test_git_init_colocate_in_git_worktree() {
 
     let worktree_dir = test_env.work_dir("worktree");
     insta::assert_snapshot!(get_colocation_status(&worktree_dir), @"
-    Workspace is currently colocated with Git.
+    Workspace 'default' is currently colocated with Git.
     Last imported/exported Git HEAD: e80a42cccd069007c7a2bb427ac7f1d10b408633
     [EOF]
     ");